Feature: added tilt stand and receiver case STLs (#65)
* feat: Added STL for switch

Switch slots into hole in the front of case. It has a slot for the switch specified in the BOM and a pincer for a top mount switch.

* IO switched to correct way around

Had them backward in first commit

* Reciever made bigger to fit

* Feat: Added a 3d printable stand

Angles are 8 degrees of tilt and 8 degrees of roll. Comfy for me, but could add others.

* Improved STL and removed keymaps

Added nice snap fittings to the receiver case and made the angle stand a lot cleaner, and sections to put non-slip feet on